862 rewrite EGF solution
